requirements.txt
fastapi
uvicorn[standard]==0.17.4
celery
flower
httpx
prefect
pydantic
#if Err: pip install -U pydantic
kombu
openai
duckduckgo-search
phidata
redis
